% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzE.R
\name{fuzzE}
\alias{fuzzE}
\title{fuzzE}
\usage{
fuzzE(x, y, caps = TRUE, rm_spec = TRUE, rm_numbers = TRUE, bump_it = FALSE)
}
\arguments{
\item{x}{First string}

\item{y}{Second string}

\item{caps}{Boolean for whether upper/lower case letters are treated the same}

\item{rm_spec}{Boolean to remove special characters}

\item{rm_numbers}{Boolean to remove numbers and suffixes (good for name matching)}

\item{bump_it}{Boolean to add rules for increasing score (in development)}
}
\value{
Score between 0 and 1 indicating strength of match
}
\description{
fuzzE
}
\examples{
fuzzE("John Wayland", "Jonathan Wayland")
# [1] 0.75
}
