% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAnswer.R
\name{showAnswer}
\alias{showAnswer}
\title{showAnswer Show the appropriate query to answer the specified question}
\usage{
showAnswer(Q = 0)
}
\arguments{
\item{Q}{The number of the question corresponding to the `questions` vector}
}
\value{
The correct query to return the desired output.
}
\description{
showAnswer Show the appropriate query to answer the specified question
}
