% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simER.R
\docType{data}
\name{simER}
\alias{simER}
\title{Simulated ER Data}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(simER)
}
\description{
Fictional healthcare dataset containing aggregated ER costs, visit counts, and severity.
}
\details{
\itemize{
 \item Cost
 \item Number of ER Visits
 \item Typical Severity Level
}
}
\examples{
data(simER)
head(simER)
}
\keyword{datasets}
