% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDB.R
\name{createDB}
\alias{createDB}
\title{createDB Creates the `moomy` database in the Global Environment}
\usage{
createDB(db = "moomy")
}
\value{

}
\description{
createDB Creates the `moomy` database in the Global Environment
}
\examples{
createDB()
}
